start lights
