+++ maps/default.rt
@@ -1,6 +1,7 @@
 A	0.2			255,255,255
 
-C	-50,0,20	0,0,0		70
+C	0,0,0	0,0,0		70
 L	-40,0,30	0.7			255,255,255
 
-sp	0,0,-1		0.5					255,0,0
+sp	0,0,-1		0.5					0,255,0
+pl 	-5,-1,-1 		0.25,0.25,-1 			255,255,225
