+++ src/ft_intersect_wrapper.c
@@ -1,7 +1,7 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   ft_intersect_wrapper.c                             :+:    :+:            */
+/*   ft_intersect_wrapper.c                             :+:      :+:    :+:   */
 /*                                                     +:+                    */
 /*   By: W2Wizard <w2.wizzard@gmail.com>              +#+                     */
 /*                                                   +#+                      */
