+++ src/lib3d/ray/intersection/ft_intersect_sphere.c
@@ -1,16 +1,17 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   ft_intersect_sphere.c                              :+:    :+:            */
+/*   ft_intersect_sphere.c                              :+:      :+:    :+:   */
 /*                                                     +:+                    */
 /*   By: lde-la-h <lde-la-h@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/04/11 17:56:15 by lde-la-h      #+#    #+#                 */
-/*   Updated: 2022/04/13 21:24:23 by W2Wizard      ########   odam.nl         */
+/*   Updated: 2022/04/14 08:33:04 by dvan-der         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "lib3d.h"
+#include "stdio.h"
 
 bool	ft_hit_sphere(t_Ray *ray, t_Sphere *sphere, t_Hit *out_hit)
 {
@@ -20,12 +21,15 @@ bool	ft_hit_sphere(t_Ray *ray, t_Sphere *sphere, t_Hit *out_hit)
 	const float		c = ft_dot_fvec3(oc, oc) - \
 						(sphere->radius * sphere->radius);
 	const float		disc = (b * b) - 4 * a * c;
-
+	
+	printf("a: %f, b: %f, c: %f, disc: %f\n", a, b, c, disc);
 	if (disc > 0)
 	{
 		out_hit->distance = (-b - sqrtf(disc)) / (2.0f * a);
+		printf("distance: %f\n", out_hit->distance);
 		return (true);
 	}
 	out_hit->distance = 0.0f;
+	printf("distance: %f\n", out_hit->distance);
 	return (false);
 }
