+++ src/rendering/ft_lights.c
@@ -1,0 +1,7 @@
+#include "MiniRT.h"
+
+void	ft_apply_lights(t_RT *rt, uint32_t *color)
+{
+	(void)rt;
+	(void)color;
+}
